+++ src/components/DownloadComponents/helpers/downloadLessonResources.tsx
@@ -3,12 +3,10 @@ import type { ResourcesToDownloadArrayType } from "../downloads.types";
 import createAndClickHiddenDownloadLink from "./createAndClickHiddenDownloadLink";
 import createDownloadResourcesLink from "./createDownloadResourcesLink";
 
-import { ViewType } from "@/common-lib/urls";
-
 const downloadLessonResources = async (
   lessonSlug: string,
   selectedResourceTypes: ResourcesToDownloadArrayType,
-  viewType: ViewType
+  isLegacyDownload: boolean,
 ) => {
   if (selectedResourceTypes?.length === 0) {
     console.log("no resources to download");
@@ -20,7 +18,7 @@ const downloadLessonResources = async (
   const downloadResourcesLink = await createDownloadResourcesLink(
     lessonSlug,
     selection,
-    viewType
+    isLegacyDownload,
   );
 
   if (downloadResourcesLink) {