+++ src/components/DownloadComponents/helpers/getDownloadResourcesExistence.tsx
@@ -1,25 +1,22 @@
-import { ViewType } from "@/common-lib/urls";
-
 const getDownloadResourcesExistence = async (
   lessonSlug: string,
   resourceTypesString: string,
-  viewType: ViewType
+  isLegacyDownload: boolean,
 ) => {
   if (!process.env.NEXT_PUBLIC_VERCEL_API_URL) {
     throw new TypeError(
-      "process.env.NEXT_PUBLIC_VERCEL_API_URL must be defined"
+      "process.env.NEXT_PUBLIC_VERCEL_API_URL must be defined",
     );
   }
   if (!process.env.NEXT_PUBLIC_DOWNLOAD_API_URL) {
     throw new TypeError(
-      "process.env.NEXT_PUBLIC_DOWNLOAD_API_URL must be defined"
+      "process.env.NEXT_PUBLIC_DOWNLOAD_API_URL must be defined",
     );
   }
 
-  const checkWhichResourcesExistEndpoint =
-    viewType === "teachers"
-      ? `${process.env.NEXT_PUBLIC_VERCEL_API_URL}/api/downloads/lesson/${lessonSlug}/check-files?selection=${resourceTypesString}`
-      : `${process.env.NEXT_PUBLIC_DOWNLOAD_API_URL}/api/lesson/${lessonSlug}/check-files?selection=${resourceTypesString}`;
+  const checkWhichResourcesExistEndpoint = isLegacyDownload
+    ? `${process.env.NEXT_PUBLIC_VERCEL_API_URL}/api/downloads/lesson/${lessonSlug}/check-files?selection=${resourceTypesString}`
+    : `${process.env.NEXT_PUBLIC_DOWNLOAD_API_URL}/api/lesson/${lessonSlug}/check-files?selection=${resourceTypesString}`;
 
   const res = await fetch(checkWhichResourcesExistEndpoint);
   const { data, error } = await res.json();