+++ src/components/EarlyReleaseExemplarUnits/EarlyReleaseExemplarUnits.test.tsx
@@ -23,7 +23,6 @@ export const earlyReleaseExemplarUnitsProps: EarlyReleaseExemplarUnitsProps = {
     occupation: "Class Teacher, St Agnes Academy",
   },
   units: units,
-  viewType: "teachers-2023",
 };
 
 const render = renderWithProviders();
@@ -33,25 +32,25 @@ describe("components/AppHeader", () => {
     render(<ExemplarUnits {...earlyReleaseExemplarUnitsProps} />);
 
     expect(
-      screen.getByText("View and download our early-release units.")
+      screen.getByText("View and download our early-release units."),
     ).toBeInTheDocument();
     expect(screen.getByText("Secondary units")).toBeInTheDocument();
     expect(
       screen.getByText(
-        "“A life saver… I didn’t think you could beat the previous Oak resources.”"
-      )
+        "“A life saver… I didn’t think you could beat the previous Oak resources.”",
+      ),
     ).toBeInTheDocument();
   });
   test("renders a list of units", () => {
     const { getByRole, getAllByRole } = render(
-      <ExemplarUnits {...earlyReleaseExemplarUnitsProps} />
+      <ExemplarUnits {...earlyReleaseExemplarUnitsProps} />,
     );
 
     const list = getByRole("list");
     expect(list).toBeInTheDocument();
     const item = getAllByRole("listitem");
     expect(item[0]).toHaveTextContent(
-      "Key stage 4 Year 10 ComputingData Representation9 lessons"
+      "Key stage 4 Year 10 ComputingData Representation9 lessons",
     );
   });
 });