+++ src/node-lib/curriculum-api/fixtures/earlyReleaseExemplarUnits.fixture.ts
@@ -6,11 +6,10 @@ type earlyReleaseExemplarUnitsFixtureProps = {
 };
 
 const earlyReleaseExemplarUnitsFixture = (
-  partial?: Partial<earlyReleaseExemplarUnitsFixtureProps>
+  partial?: Partial<earlyReleaseExemplarUnitsFixtureProps>,
 ): earlyReleaseExemplarUnitsFixtureProps => {
   return {
     primary: {
-      viewType: "teachers-2023",
       subjectIconBackground: "lavender",
       heading: "Primary units",
       subHeading: "View and download our early-release units.",
@@ -215,7 +214,6 @@ const earlyReleaseExemplarUnitsFixture = (
       ],
     },
     secondary: {
-      viewType: "teachers-2023",
       subjectIconBackground: "pink",
       heading: "Secondary units",
       subHeading: "View and download our early-release units.",